Improve test structure
Improve test code structure following
https://github.com/operator-framework/operator-marketplace/blob/master/docs/e2e-testing.md
