+++ test/helpers/helpers.go
@@ -1,0 +1,11 @@
+package helpers
+
+import "testing"
+
+// AssertNoError confirms the error returned is nil
+func AssertNoError(t *testing.T, err error) {
+	t.Helper()
+	if err != nil {
+		t.Fatal(err)
+	}
+}
